Fixing DB alter SQL issues with MatchAnalysis Plugin

+++ application/modules/editor/Plugins/MatchAnalysis/database/deinstall_004-editorPlugin-MatchAnalysis-mysql.sql
@@ -31,28 +31,12 @@ DROP TABLE IF EXISTS `LEK_match_analysis_taskassoc`;
 DELETE FROM `Zf_worker_dependencies` where `worker` =  'editor_Plugins_MatchAnalysis_Worker' AND `dependency` = 'editor_Models_Import_Worker';
 DELETE FROM `Zf_worker_dependencies` where `worker` =  'editor_Plugins_TermTagger_Worker_TermTaggerImport' AND `dependency` = 'editor_Plugins_MatchAnalysis_Worker';
 DELETE FROM `Zf_worker_dependencies` where `worker` =  'editor_Models_Import_Worker_SetTaskToOpen' AND `dependency` = 'editor_Plugins_MatchAnalysis_Worker';
+DELETE FROM `Zf_worker_dependencies` WHERE `worker` = 'editor_Plugins_GlobalesePreTranslation_Worker' AND `dependency` = 'editor_Plugins_MatchAnalysis_Worker';
 
 DELETE FROM `Zf_configuration` 
 WHERE `name` = 'runtimeOptions.worker.editor_Plugins_MatchAnalysis_Worker.maxParallelWorkers';
 
-DELETE FROM `Zf_acl_rules` 
-WHERE `right` = 'pluginMatchAnalysisMatchAnalysis' 
-AND `resource` = 'frontend' 
-AND `role` = 'pm'
-AND `module` = 'editor';
+DELETE FROM `Zf_acl_rules` WHERE `right` = 'pluginMatchAnalysisMatchAnalysis' AND `resource` = 'frontend' AND `role` = 'pm' AND `module` = 'editor';
+DELETE FROM `Zf_acl_rules` WHERE `right` = 'all' AND `resource` = 'editor_plugins_matchanalysis_matchanalysis' AND `role` = 'pm' AND `module` = 'editor';
+DELETE FROM `Zf_acl_rules` WHERE `right` = 'editorAnalysisTask' AND `resource` = 'frontend' AND `role` = 'pm' AND `module` = 'editor';
 
-DELETE FROM `Zf_acl_rules` 
-WHERE `right` = 'all' 
-AND `resource` = 'editor_plugins_matchanalysis_matchanalysis' 
-AND `role` = 'pm'
-AND `module` = 'editor';
-
-DELETE FROM `Zf_acl_rules` 
-WHERE `right` = 'editorAnalysisTask' 
-AND `resource` = 'frontend' 
-AND `role` = 'pm'
-AND `module` = 'editor';
-
-DELETE FROM `Zf_worker_dependencies` 
-WHERE `worker` = 'editor_Plugins_GlobalesePreTranslation_Worker' 
-AND `dependency` = 'editor_Plugins_MatchAnalysis_Worker';
